Refactor `NFT collection overview` module to use INftProvider
