+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/market/overview/MarketOverviewScreen.kt
@@ -27,7 +27,6 @@ import io.horizontalsystems.bankwallet.modules.market.platform.MarketPlatformFra
 import io.horizontalsystems.bankwallet.modules.market.topcoins.MarketTopCoinsFragment
 import io.horizontalsystems.bankwallet.modules.market.topnftcollections.TopNftCollectionsFragment
 import io.horizontalsystems.bankwallet.modules.market.topplatforms.TopPlatformsFragment
-import io.horizontalsystems.bankwallet.modules.nft.collection.NftCollectionFragment
 import io.horizontalsystems.bankwallet.ui.compose.HSSwipeRefresh
 import io.horizontalsystems.bankwallet.ui.compose.components.ListErrorView
 
@@ -105,8 +104,8 @@ fun MarketOverviewScreen(
                                     viewModel.onSelectTopNftsTimeDuration(timeDuration)
                                 },
                                 onClickCollection = { collectionUid ->
-                                    val args = NftCollectionFragment.prepareParams(collectionUid)
-                                    navController.slideFromBottom(R.id.nftCollectionFragment, args)
+//                                    val args = NftCollectionFragment.prepareParams(collectionUid)
+//                                    navController.slideFromBottom(R.id.nftCollectionFragment, args)
                                 },
                                 onClickSeeAll = {
                                     val (sortingField, timeDuration) = viewModel.topNftCollectionsParams
