+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/market/topnftcollections/TopNftCollectionsFragment.kt
@@ -26,13 +26,11 @@ import androidx.fragment.app.viewModels
 import com.google.accompanist.swiperefresh.rememberSwipeRefreshState
 import io.horizontalsystems.bankwallet.R
 import io.horizontalsystems.bankwallet.core.BaseFragment
-import io.horizontalsystems.bankwallet.core.slideFromBottom
 import io.horizontalsystems.bankwallet.entities.ViewState
 import io.horizontalsystems.bankwallet.modules.coin.overview.Loading
 import io.horizontalsystems.bankwallet.modules.market.MarketDataValue
 import io.horizontalsystems.bankwallet.modules.market.SortingField
 import io.horizontalsystems.bankwallet.modules.market.TimeDuration
-import io.horizontalsystems.bankwallet.modules.nft.collection.NftCollectionFragment
 import io.horizontalsystems.bankwallet.ui.compose.ComposeAppTheme
 import io.horizontalsystems.bankwallet.ui.compose.HSSwipeRefresh
 import io.horizontalsystems.bankwallet.ui.compose.components.*
@@ -65,8 +63,8 @@ class TopNftCollectionsFragment : BaseFragment() {
                         viewModel,
                         { findNavController().popBackStack() },
                         { collectionUid ->
-                            val args = NftCollectionFragment.prepareParams(collectionUid)
-                            findNavController().slideFromBottom(R.id.nftCollectionFragment, args)
+//                            val args = NftCollectionFragment.prepareParams(collectionUid)
+//                            findNavController().slideFromBottom(R.id.nftCollectionFragment, args)
                         }
                     )
                 }
