+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/nft/collection/NftCollectionModule.kt
@@ -5,15 +5,21 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import io.horizontalsystems.bankwallet.R
 import io.horizontalsystems.bankwallet.core.App
+import io.horizontalsystems.bankwallet.modules.nft.collection.overview.NftCollectionOverviewService
+import io.horizontalsystems.bankwallet.modules.nft.collection.overview.NftCollectionViewModel
 import io.horizontalsystems.bankwallet.modules.xrate.XRateService
+import io.horizontalsystems.marketkit.models.BlockchainType
 
 object NftCollectionModule {
 
-    class Factory(private val collectionUid: String) : ViewModelProvider.Factory {
+    class Factory(
+        private val blockchainType: BlockchainType,
+        private val collectionUid: String
+    ) : ViewModelProvider.Factory {
 
         @Suppress("UNCHECKED_CAST")
         override fun <T : ViewModel> create(modelClass: Class<T>): T {
-            val service = NftCollectionService(collectionUid, App.marketKit)
+            val service = NftCollectionOverviewService(blockchainType, collectionUid, App.nftMetadataManager.provider(blockchainType), App.marketKit)
             return NftCollectionViewModel(
                 service,
                 App.numberFormatter,
