+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/nft/collection/overview/NftCollectionOverviewService.kt
@@ -1,7 +1,10 @@
-package io.horizontalsystems.bankwallet.modules.nft.collection
+package io.horizontalsystems.bankwallet.modules.nft.collection.overview
 
+import io.horizontalsystems.bankwallet.core.adapters.nft.INftProvider
 import io.horizontalsystems.bankwallet.core.managers.MarketKitWrapper
-import io.horizontalsystems.marketkit.models.NftCollection
+import io.horizontalsystems.bankwallet.entities.nft.NftCollectionMetadata
+import io.horizontalsystems.marketkit.models.Blockchain
+import io.horizontalsystems.marketkit.models.BlockchainType
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -9,14 +12,22 @@ import kotlinx.coroutines.flow.filterNotNull
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
 
-class NftCollectionService(
-    val collectionUid: String,
+class NftCollectionOverviewService(
+    val blockchainType: BlockchainType,
+    val providerCollectionUid: String,
+    private val provider: INftProvider,
     private val marketKit: MarketKitWrapper
 ) {
     private var fetchingJob: Job? = null
 
-    private val _nftCollection = MutableStateFlow<Result<NftCollection>?>(null)
+    private val _nftCollection = MutableStateFlow<Result<NftCollectionMetadata>?>(null)
     val nftCollection = _nftCollection.filterNotNull()
+
+    val providerTitle = provider.title
+    val providerIcon = provider.icon
+
+    val blockchain: Blockchain?
+        get() = marketKit.blockchain(blockchainType.uid)
 
     suspend fun start() {
         fetch()
@@ -31,7 +42,7 @@ class NftCollectionService(
 
         fetchingJob = launch {
             try {
-                val collection = marketKit.nftCollection(collectionUid)
+                val collection = provider.collectionMetadata(blockchainType, providerCollectionUid)
 
                 _nftCollection.emit(Result.success(collection))
             } catch (error: Exception) {
